Let us always disable remote account creation.
When the automatic configuration wizard is enabled we probably want to
use *it*, and not the remote account creation feature. If that is
something TorBirdys want to make enabled as an option, it should not
be baked into this same pref.

However, it probably is something TorBirdy should not encourage (at
least we in Tails decided to disable it), so always hiding it seems
like a good idea.

This fixes Tails bug #10464:

    https://labs.riseup.net/code/issues/10464