Set authentication for Gmail to OAuth2 (#17118)
For Gmail, the authentication type is set to OAuth2 instead of the default
"normal password" which is what most providers use (Google doesn't allow
non-OAuth2 logins). To make this work with the manual configuration wizard in
TorBirdy, the authorization step takes place when the user checks the account
for the first time, and is not done automatically. We can probably improve
this but it's better to wait for the patches that secure the automatic
configuration wizard to be merged upstream.  See bugzilla 669238.