Reflect TorBirdy's state in UI after uninstall/disable (#10683)
When TorBirdy is uninstalled or disabled, the status bar text and colour is
updated to reflect TorBirdy's current state. Prior to this commit, if a user
uninstalled/disabled TorBirdy, the status bar text would still say "TorBirdy
Enabled". Now the status bar is updated as TorBirdy's state changes to avoid
confusing the user.